Refactor fitness component imports and restructure utility functions
- Updated import paths for fitness-related components and utilities to improve organization.
- Moved `ExerciseHistory` and `WorkoutRevisionScheduledAlert` imports to their respective new locations.
- Consolidated utility functions into `utils.tsx` for better maintainability.
- Removed the deprecated `fitness.tsx` file to streamline the codebase.
- Enhanced the `display-items.tsx` file with additional display components for exercises and workouts.
