+++ apps/frontend/app/components/common/summary.tsx
@@ -33,7 +33,7 @@ import {
 	useUserUnitSystem,
 } from "~/lib/shared/hooks";
 import { MediaColors, getMetadataIcon } from "~/lib/shared/media-utils";
-import { displayWeightWithUnit } from "../fitness";
+import { displayWeightWithUnit } from "../fitness/utils";
 
 export const DisplaySummarySection = ({
 	latestUserSummary,
